+++ projects/js-packages/charts/src/hooks/index.ts
@@ -9,3 +9,4 @@ export { useTextTruncation } from './use-text-truncation';
 export { useZeroValueDisplay } from './use-zero-value-display';
 export { useInteractiveLegendData } from './use-interactive-legend-data';
 export { usePrefersReducedMotion } from './use-prefers-reduced-motion';
+export { useTooltipPortalRelocator } from './use-tooltip-portal-relocator';