+++ projects/js-packages/api/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### This is a list detailing changes for the Jetpack RNA Components package releases.
 
+## [1.0.17] - 2026-02-18
+### Changed
+- Set `.repository.url` in `package.json` to the mirror repo rather than the monorepo. [#47149]
+
 ## [1.0.16] - 2026-02-02
 ### Changed
 - Update package dependencies. [#46854]
@@ -486,6 +490,7 @@
 - Add the API methods left behind by the previous PR.
 - Initial release of jetpack-api package
 
+[1.0.17]: https://github.com/Automattic/jetpack-api/compare/v1.0.16...v1.0.17
 [1.0.16]: https://github.com/Automattic/jetpack-api/compare/v1.0.15...v1.0.16
 [1.0.15]: https://github.com/Automattic/jetpack-api/compare/v1.0.14...v1.0.15
 [1.0.14]: https://github.com/Automattic/jetpack-api/compare/v1.0.13...v1.0.14