+++ projects/js-packages/base-styles/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.17] - 2026-02-18
+### Changed
+- Set `.repository.url` in `package.json` to the mirror repo rather than the monorepo. [#47149]
+
 ## [1.0.16] - 2026-02-16
 ### Fixed
 - Compatibility: Clean up deprecated CSS. [#47067]
@@ -468,6 +472,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated package dependencies.
 - Use Node 16.7.0 in tooling. This shouldn't change the behavior of the code itself.
 
+[1.0.17]: https://github.com/Automattic/jetpack-base-styles/compare/1.0.16...1.0.17
 [1.0.16]: https://github.com/Automattic/jetpack-base-styles/compare/1.0.15...1.0.16
 [1.0.15]: https://github.com/Automattic/jetpack-base-styles/compare/1.0.14...1.0.15
 [1.0.14]: https://github.com/Automattic/jetpack-base-styles/compare/1.0.13...1.0.14